requirements.txt
# API et serveur
fastapi>=0.103.0
uvicorn>=0.23.0
pydantic>=2.3.0
python-multipart>=0.0.6
email-validator>=2.0.0

# Base de données
sqlalchemy>=2.0.0
alembic>=1.11.0
psycopg2-binary>=2.9.5

# Authentification et sécurité
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
pyjwt>=2.7.0

# Outils de base
numpy>=1.26.4
pandas>=2.0.0
tenacity>=8.2.2
psutil>=5.9.5
pillow>=10.0.0
matplotlib>=3.7.0

# Traitement video et audio
opencv-python>=4.7.0
ffmpeg-python>=0.2.0
imageio>=2.25.0
moviepy>=1.0.3
pyannote.audio>=2.1.1
librosa>=0.10.0

# Modèles et IA
# Conserve la version spécifiée pour vllm
vllm>=0.7.3; platform_system != "Darwin" or platform_machine != "arm64"
torch>=2.0.0
transformers==4.49.0  # Version spécifique pour Huggingface
datasets>=2.12.0
accelerate>=0.20.0
optimum>=1.8.0
huggingface-hub>=0.16.4
tokenizers>=0.13.3
jinja2>=3.1.2
openai>=0.27.0  # Pour la compatibilité API
deepspeed>=0.9.5; platform_system != "Darwin"
einops>=0.6.1  # Pour les transformers visuels

# Segmentation
spacy>=3.6.1
sentence-transformers>=2.2.2
scikit-learn>=1.3.0

# Modèles linguistiques français
fr-core-news-sm@ https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.6.0/fr_core_news_sm-3.6.0-py3-none-any.whl

# Utilitaires
pyyaml>=6.0.1
aiofiles>=23.1.0
async-timeout>=4.0.3
tqdm>=4.65.0
retry>=0.9.2
colorama>=0.4.6
python-dotenv>=1.0.0
pymongo>=4.3.3  # Pour la compatibilité avec certains modèles
httpx>=0.24.1  # Pour les requêtes HTTP asynchrones

# Paiements
stripe>=5.4.0

# Tests
pytest>=7.3.1
pytest-asyncio>=0.21.0
httpx>=0.24.1

# Outils de développement
black>=23.3.0
isort>=5.12.0
mypy>=1.3.0
flake8>=6.0.0

# Monitoring
prometheus-client>=0.17.0

langdetect>=1.0.9
sacremoses>=0.0.53
sentencepiece>=0.1.97